requirements.txt
flask>=3.0.0
litellm>=1.0.0
anthropic>=0.30.0
google-genai>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
